Download linux-metadata tarball
Previously, the entire threadbare repository was being cloned just to
fetch a few static files. Apart from the fact that it's huge, it uses
Git LFS; while this worked for the initial build, fetching it
unauthenticated is now failing on Flathub (interesting and potentially
bad). It also relies on a patched copy of flatpak-builder.

I added a step to our release process to make a tarball of the small
number of files needed here. Update the manifest to use this. Drop the section about Git LFS from the README.